move Guest and User to components folder
