+++ src/checkout/components/UserAddressSelector/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UserAddressSelector";
