+++ src/components/index.ts
@@ -60,3 +60,4 @@ export {
 } from "./Overlay";
 export { default as Select } from "./Select";
 export { default as Modal } from "./Modal";
+export { default as Error } from "./Error";
